Refactor things into files
