+++ docs/peewee_async/api_older.rst
@@ -6,12 +6,12 @@ Low-level (older) API
 Select, update, delete
 ----------------------
 
-.. autofunction:: peewee_async.execute
-.. autofunction:: peewee_async.get_object
-.. autofunction:: peewee_async.create_object
-.. autofunction:: peewee_async.delete_object
-.. autofunction:: peewee_async.update_object
-.. autofunction:: peewee_async.prefetch
+.. autofunction:: src.queries.execute
+.. autofunction:: src.queries.get_object
+.. autofunction:: src.queries.create_object
+.. autofunction:: src.queries.delete_object
+.. autofunction:: src.queries.update_object
+.. autofunction:: src.queries.prefetch
 
 Transactions
 ------------
@@ -21,24 +21,24 @@ Transactions required Python 3.5+ to work, because their syntax is based on asyn
 **Important note** transactions rely on data isolation on `asyncio` per-task basis.
 That means, all queries for single transaction should be performed **within same task**.
 
-.. autofunction:: peewee_async.atomic
-.. autofunction:: peewee_async.savepoint
-.. autofunction:: peewee_async.transaction
+.. autofunction:: src.transactions.atomic
+.. autofunction:: src.transactions.savepoint
+.. autofunction:: src.transactions.transaction
 
 Aggregation
 -----------
 
-.. autofunction:: peewee_async.count
-.. autofunction:: peewee_async.scalar
+.. autofunction:: src.queries.count
+.. autofunction:: src.queries.scalar
 
 Databases
 ---------
 
-.. autoclass:: peewee_async.PostgresqlDatabase
+.. autoclass:: src.postgres.PostgresqlDatabase
     :members: connect_async, atomic_async, transaction_async, savepoint_async
     :noindex:
 
-.. autoclass:: peewee_async.PooledPostgresqlDatabase
+.. autoclass:: src.postgres.PooledPostgresqlDatabase
     :members: connect_async
     :noindex:
